Extract each parser (binary & text) into their own modules
